+++ Classes/00-Base/DISMPackageDefinition.ps1
@@ -1,0 +1,11 @@
+class DISMPackageDefinition : Package {
+    [string]$ServicingType
+    [string]$State
+    [string]$SourcePath
+
+    DISMPackageDefinition() {
+        $this.Provider = 'DISM'
+        $this.Source = 'Online Windows Image'
+        $this.RequiresElevation = $true
+    }
+}